test: user service
